Version 1.1.4 released
> Update to TinyMCE 4.0.5
The main modification is a new option to automatically restore the
TinyMCE draft when there is one and when the editor is empty. This
option can be configured inside the XenForo Options (TinyMCE Quattro
section of course)
> Fix a visual bug with the right border of the quickreply editor with
Chrome
> Fix some typo with phrases that prevented to translate some phrases
(Thanks to HWS)
>Patch for the autoresize plugin that was not lauched insided the
XenForo Overlay with some browsers (IE, Chrome ; was working with FF)
(Thanks to HWS)
> Fix for the smilies descriptions that were not displayed (Thanks to
HWS)

It seems the patches for TinyMCE/JqueryTools Overlay/Firefox  are no
more needed. It now works. I will wait a little before deleting them.