Version 1.5.0 released
-Style integration to XenForo entirely rewritten
#It uses now the Play With Colors! addon to automatically adapt to your
theme and to reduce drastically the number of options. It has been
tested on several personal themes and the integration was almost correct
every time. If you haven't installed "Play With Colors", don't worry the
integration is automatic.
#If this new integration doesn't suite your theme, you can still use any
of TinyMCE original themes or create your own using the TinyMCE Skin
Creator.
-The problem with the lists inside a table should have been fixed.
(Reported by @ZeZeene)
-The copy/upload image function has been disabled since it was only
working with Firefox . I'm waiting for an answer from the TinyMCE
developers.