Version 2.6.3 WIP
- Fix a bug with the mini parser when the same opening parsing character
was used several times ; ie: [[b]This is[/b] a test].
- Mce detection modified. How it worked before: to detect if MCE is
enabled, the CCV ($controllerName, $controllerAction, $viewName) are
checked. If a user loads a new page, then these values will change. If
the user comes back to the original page, it is unlikely that the MCE
will be detected as enabled. This was a problem when the user came back
and toggled between the RTE<=>Bb Code editors (some layouts of the
WYSIWYG Bb Codes could be lost) . A workaround has been found.